Add mix task for adding documentation
